gui/source/requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==25.2.10
gitdb==4.0.12
GitPython==3.1.45
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==2.2.0
numpy==1.26.4
nvidia-cublas-cu12==12.9.1.4
nvidia-cudnn-cu12==9.13.0.50
onnxruntime-directml==1.22.0
onnxruntime-gpu==1.22.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pydeck==0.9.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
